refactored tech tag customizer to file
